+++ app/src/main/java/com/yasn/purchase/model/HomeModel.java
@@ -204,7 +204,7 @@ public class HomeModel implements Serializable{
          * priceDisplayType : 0
          */
 
-        private int seePrice;
+        private double seePrice;
         private long endDate;
         private int level_id;
         private String employee_auth;
@@ -252,11 +252,11 @@ public class HomeModel implements Serializable{
             this.provinceId = provinceId;
         }
 
-        public int getSeePrice() {
+        public double getSeePrice() {
             return seePrice;
         }
 
-        public void setSeePrice(int seePrice) {
+        public void setSeePrice(double seePrice) {
             this.seePrice = seePrice;
         }
